move sin to functions
