+++ decedric/core.py
@@ -226,19 +226,6 @@ class Pow(Function):
 
 def pow(x, c):
     return Pow(c)(x)
-
-class Sin(Function):
-    def forward(self, x):
-        y = np.sin(x)
-        return y
-    
-    def backward(self, gy):
-        x = self.inputs[0].data
-        gx = gy * np.cos(x)
-        return gx
-
-def sin(x):
-    return Sin()(x)
     
 @contextlib.contextmanager
 def using_config(name, value):
