+++ main.py
@@ -3,18 +3,10 @@ if '__file__' in globals():
     sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
 import numpy as np
+import decedric.functions as F
 from decedric import Variable
 
-def f(x):
-    y = x ** 4 - 2 * x ** 2
-    return y
-
-x = Variable(np.array(2.0))
-y = f(x)
-y.backward(create_graph=True)
-print(x.grad)
-
-gx = x.grad
-x.cleargrad()
-gx.backward()
-print(x.grad)
+x = Variable(np.array([[1,2,3],[4,5,6]]))
+c = Variable(np.array([[10,20,30],[40,50,60]]))
+y = x+c
+print(y)
